src/apply.rs: apply timestamps from previous deployments
This is a key piece of the "layered metadata" framework. Once we've
applied a timestamp to a post, we need to re-apply it in the future! We
accomplish this by scanning the content of the `deploy` branch for dates
that should be updated. It turns out to be pretty straightforward.